pyproject.toml
[tool.poetry]
name = "linearproj"
version = "0.1.0"
description = ""
authors = ["Harikaran Subbaraj <hari.subbaraj@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.1"
requests = "^2.31.0"
openai = "^1.13.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
